+++ tests/tmc/test_assign_release_command_low.py
@@ -320,7 +320,7 @@ def test_release_exception_propagation(
         central_node_low.central_node,
         [exception_message],
         unique_id[0],
-        ResultCode.REJECTED,
+        ResultCode.NOT_ALLOWED,
     )
     exception_message = "Timeout has occurred, command failed"